requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
scapy==2.5.0
firebase-admin==6.3.0
python-dotenv==1.0.0
